Move Android dev folder to Atlas/android for uniform project structure.
