+++ Atlas/android/.gitignore
