+++ Atlas/android/.idea/copyright/profiles_settings.xml
