+++ Atlas/android/app/.gitignore
