+++ Atlas/android/gradle.properties
