+++ Atlas/android/settings.gradle
